otel: add global logger provider
