+++ bridge/opencensus/test/go.mod
@@ -17,6 +17,7 @@ require (
 	github.com/golang/groupcache v0.0.0-20241129210726-2c02b8208cf8 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	go.opentelemetry.io/auto/sdk v1.2.1 // indirect
+	go.opentelemetry.io/otel/log v0.22.0 // indirect
 	go.opentelemetry.io/otel/metric v1.46.0 // indirect
 	go.opentelemetry.io/otel/sdk/metric v1.46.0 // indirect
 	golang.org/x/sys v0.47.0 // indirect
@@ -35,3 +36,5 @@ replace go.opentelemetry.io/otel/metric => ../../../metric
 replace go.opentelemetry.io/otel/sdk/metric => ../../../sdk/metric
 
 replace go.opentelemetry.io/otel/metric/x => ../../../metric/x
+
+replace go.opentelemetry.io/otel/log => ../../../log
