+++ exporters/otlp/otlpmetric/otlpmetricgrpc/go.mod
@@ -25,6 +25,7 @@ require (
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/grpc-ecosystem/grpc-gateway/v2 v2.30.0 // indirect
 	go.opentelemetry.io/auto/sdk v1.2.1 // indirect
+	go.opentelemetry.io/otel/log v0.22.0 // indirect
 	go.opentelemetry.io/otel/trace v1.46.0 // indirect
 	go.yaml.in/yaml/v3 v3.0.5 // indirect
 	golang.org/x/net v0.58.0 // indirect
@@ -44,3 +45,5 @@ replace go.opentelemetry.io/otel/metric => ../../../../metric
 replace go.opentelemetry.io/otel/trace => ../../../../trace
 
 replace go.opentelemetry.io/otel/metric/x => ../../../../metric/x
+
+replace go.opentelemetry.io/otel/log => ../../../../log
