+++ exporters/stdout/stdoutlog/example_test.go
@@ -6,8 +6,8 @@ package stdoutlog_test
 import (
 	"context"
 
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/exporters/stdout/stdoutlog"
-	"go.opentelemetry.io/otel/log/global"
 	"go.opentelemetry.io/otel/sdk/log"
 )
 
@@ -25,7 +25,7 @@ func Example() {
 		}
 	}()
 
-	global.SetLoggerProvider(provider)
+	otel.SetLoggerProvider(provider)
 
 	// From here, the provider can be used by instrumentation to collect
 	// telemetry.
