+++ exporters/stdout/stdouttrace/go.mod
@@ -22,6 +22,7 @@ require (
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	go.opentelemetry.io/auto/sdk v1.2.1 // indirect
+	go.opentelemetry.io/otel/log v0.22.0 // indirect
 	go.yaml.in/yaml/v3 v3.0.5 // indirect
 	golang.org/x/sys v0.47.0 // indirect
 )
@@ -33,3 +34,5 @@ replace go.opentelemetry.io/otel/metric => ../../../metric
 replace go.opentelemetry.io/otel/sdk/metric => ../../../sdk/metric
 
 replace go.opentelemetry.io/otel/metric/x => ../../../metric/x
+
+replace go.opentelemetry.io/otel/log => ../../../log
