+++ go.mod
@@ -9,6 +9,7 @@ require (
 	github.com/google/go-cmp v0.7.0
 	github.com/stretchr/testify v1.12.1
 	go.opentelemetry.io/auto/sdk v1.2.1
+	go.opentelemetry.io/otel/log v0.22.0
 	go.opentelemetry.io/otel/metric v1.46.0
 	go.opentelemetry.io/otel/trace v1.46.0
 )
@@ -18,3 +19,5 @@ require go.yaml.in/yaml/v3 v3.0.5 // indirect
 replace go.opentelemetry.io/otel/trace => ./trace
 
 replace go.opentelemetry.io/otel/metric => ./metric
+
+replace go.opentelemetry.io/otel/log => ./log
