+++ internal/global/logger.go
@@ -1,8 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-// Package global is the internal implementation of the OpenTelemetry global
-// Logs API.
 package global
 
 import (
