+++ log/global/README.md
@@ -1,3 +1,5 @@
 # Log Global
 
+> Deprecated: Use the equivalent APIs in [go.opentelemetry.io/otel](https://pkg.go.dev/go.opentelemetry.io/otel).
+
 [![PkgGoDev](https://pkg.go.dev/badge/go.opentelemetry.io/otel/log/global)](https://pkg.go.dev/go.opentelemetry.io/otel/log/global)
