+++ log/go.mod
@@ -3,13 +3,13 @@ module go.opentelemetry.io/otel/log
 go 1.25.0
 
 require (
-	github.com/go-logr/logr v1.4.4
 	github.com/stretchr/testify v1.12.1
 	go.opentelemetry.io/otel v1.46.0
 )
 
 require (
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/go-logr/logr v1.4.4 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	go.opentelemetry.io/auto/sdk v1.2.1 // indirect
 	go.opentelemetry.io/otel/metric v1.46.0 // indirect
