+++ metric/go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/go-logr/logr v1.4.4 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	go.opentelemetry.io/auto/sdk v1.2.1 // indirect
+	go.opentelemetry.io/otel/log v0.22.0 // indirect
 	go.opentelemetry.io/otel/trace v1.46.0 // indirect
 	go.yaml.in/yaml/v3 v3.0.5 // indirect
 )
@@ -19,3 +20,5 @@ require (
 replace go.opentelemetry.io/otel => ../
 
 replace go.opentelemetry.io/otel/trace => ../trace
+
+replace go.opentelemetry.io/otel/log => ../log
