+++ metric/x/go.mod
@@ -14,3 +14,5 @@ replace go.opentelemetry.io/otel/metric => ../../metric
 replace go.opentelemetry.io/otel/trace => ../../trace
 
 replace go.opentelemetry.io/otel => ../..
+
+replace go.opentelemetry.io/otel/log => ../../log
