+++ trace/go.mod
@@ -16,3 +16,5 @@ require (
 )
 
 replace go.opentelemetry.io/otel/metric => ../metric
+
+replace go.opentelemetry.io/otel/log => ../log
