+++ trace/internal/telemetry/test/go.mod
@@ -23,3 +23,5 @@ replace go.opentelemetry.io/otel/trace => ../../..
 replace go.opentelemetry.io/otel => ../../../..
 
 replace go.opentelemetry.io/otel/metric => ../../../../metric
+
+replace go.opentelemetry.io/otel/log => ../../../../log
